requirements.txt
git+https://github.com/Nix-da/hopfield-layers.git
-e ./worm_world

gym~=0.26.0
gymnasium~=0.29.1
setuptools~=68.2.2
pygame~=2.1.0
numpy~=1.25.2
matplotlib~=3.8.0
seaborn~=0.12.2
tqdm~=4.66.1
torch>2.0.0+cu121
imageio~=2.31.6
rl_zoo3~=2.2.1
rliable
swig
cmake
ffmpeg
tensorboard
moviepy
mysqlclient
plotly
scikit-learn
kaleido

pandas~=2.1.3
optuna~=3.4.0
SQLAlchemy~=2.0.23
PyYAML~=6.0.1

statsmodels~=0.14.0
scipy~=1.13.0
autograd~=1.6.2
worms~=0.1